create template files
