+++ packages/bigframes/scripts/templates/operation.py.j2
@@ -1,0 +1,46 @@
+{% include 'license.py.j2' %}
+
+#
+# DO NOT MODIFY THIS FILE DIRECTLY.
+# This file was generated from: {{ yaml_path }}
+# by the script: {{ script_path }}
+
+from __future__ import annotations
+
+import datetime
+from typing import Any, Literal, Optional, TypeVar, Union
+
+from bigframes import dtypes
+import bigframes.bigquery._googlesql
+import bigframes.core.col
+import bigframes.core.expression as ex
+import bigframes.core.sentinels as sentinels
+from bigframes.operations import googlesql
+import bigframes.operations as ops
+import bigframes.series as series
+
+T = TypeVar("T", series.Series, bigframes.core.col.Expression)
+
+{% for op in ops %}
+{{ op.internal_name }} = googlesql.GoogleSqlScalarOp(
+    "{{ op.sql_name }}",
+    args=({{ op.arg_specs }}),
+    signature={{ op.signature }},
+)
+{% endfor %}
+{% for func in functions %}
+
+
+def {{ func.name }}(
+{% for arg in func.args %}
+    {{ arg.name }}: Union[T, bigframes.core.col.Expression, {{ arg.type_hint }}]{% if arg.default %} = {{ arg.default }}{% endif %},
+{% endfor %}
+) -> T:
+    """{{ func.description }}"""
+    return bigframes.bigquery._googlesql.apply_googlesql_scalar_op(
+        {{ func.op_name }},
+{% for arg in func.args %}
+        {{ arg.name }},
+{% endfor %}
+    )  # type: ignore
+{% endfor %}
