+++ packages/bigframes/tests/unit/bigquery/_operations/snapshots/test_aead/test_encrypt/out.sql
@@ -1,4 +1,4 @@
 SELECT
   `rowindex`,
-  AEAD.ENCRYPT(`bytes_col`, `string_col`, `string_col`) AS `0`
+  AEAD.ENCRYPT(`bytes_col`, `bytes_col`, `bytes_col`) AS `0`
 FROM `bigframes-dev`.`sqlglot_test`.`scalar_types` AS `bft_0`
